ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.50.2
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
qdrant-client==1.15.1
pypdf==4.3.1
